refactor(suite): merge modal to index file to simplify structure
