+++ packages/suite/src/components/suite/modals/AddAccount/index.tsx
@@ -1,6 +1,7 @@
 import React, { useState } from 'react';
-import { Button } from '@trezor/components';
-import { Translation } from '@suite-components';
+import styled from 'styled-components';
+import { Button, P, variables } from '@trezor/components';
+import { Translation, Modal } from '@suite-components';
 import { NETWORKS } from '@wallet-config';
 import { Account, Network } from '@wallet-types';
 import { TrezorDevice } from '@suite-types';
@@ -10,11 +11,33 @@ import * as walletSettingsActions from '@settings-actions/walletSettingsActions'
 import * as routerActions from '@suite-actions/routerActions';
 import { partition } from '@suite-utils/array';
 
+import AccountTypeSelect from './components/AccountTypeSelect';
+import AddAccount from './components/AddAccount';
+import EnableNetwork from './components/EnableNetwork';
 import NetworkUnavailable from './components/NetworkUnavailable';
 import NetworkInternal from './components/NetworkInternal';
 import AddAccountButton from './components/AddAccountButton';
-import Wrapper from './components/Wrapper';
 
+const StyledModal = styled(props => <Modal {...props} />)`
+    min-height: 550px;
+    text-align: left;
+`;
+
+const Actions = styled.div`
+    display: flex;
+    justify-content: center;
+`;
+
+const Title = styled(P)`
+    margin-right: 9px;
+    padding: 14px 0%;
+    font-weight: ${variables.FONT_WEIGHT.MEDIUM};
+`;
+
+const Expander = styled.div`
+    display: flex;
+    flex: 1;
+`;
 interface Props {
     device: TrezorDevice;
     onCancel: () => void;
@@ -22,7 +45,7 @@ interface Props {
     noRedirect?: boolean;
 }
 
-const AddAccount = (props: Props) => {
+const AddAccountModal = ({ device, onCancel, symbol, noRedirect }: Props) => {
     const { changeAccountVisibility, changeCoinVisibility, goto } = useActions({
         changeAccountVisibility: accountActions.changeAccountVisibility,
         changeCoinVisibility: walletSettingsActions.changeCoinVisibility,
@@ -38,18 +61,19 @@ const AddAccount = (props: Props) => {
     const internalNetworks = NETWORKS.filter(n => !n.accountType && !n.isHidden);
 
     // Collect device unavailable capabilities
-    const unavailableCapabilities = props.device.features
-        ? props.device.unavailableCapabilities
-        : {};
+    const unavailableCapabilities = device.features ? device.unavailableCapabilities : {};
 
     // applied only when changing account in coinmarket exchange receive options context so far
-    const preselectedNetwork =
-        props.symbol && internalNetworks.find(n => n.symbol === props.symbol);
-    const networkPinned = !!props.symbol;
+    const networkPinned = !!symbol;
+    const preselectedNetwork = symbol && internalNetworks.find(n => n.symbol === symbol);
 
     const [network, setNetwork] = useState<Network | undefined>(preselectedNetwork);
 
     const networkEnabled = !!network && enabledNetworksSymbols.includes(network.symbol);
+
+    // Check device capabilities
+    // Display: unavailable network
+    const unavailableCapability = network && unavailableCapabilities[network.symbol];
 
     const [enabledNetworks, disabledNetworks] = partition(internalNetworks, network =>
         enabledNetworksSymbols.includes(network.symbol),
@@ -70,37 +94,11 @@ const AddAccount = (props: Props) => {
         }
     };
 
-    // Common props for Wrapper
-    const wrapperProps = {
-        network,
-        networkEnabled,
-        networkPinned,
-        enabledNetworks,
-        disabledMainnetNetworks,
-        disabledTestnetNetworks,
-        handleNetworkSelection,
-        handleAccountTypeSelection: setNetwork,
-        onCancel: props.onCancel,
-        unavailableCapabilities,
-    };
-
-    // Check device capabilities
-    // Display: unavailable network
-    const unavailableCapability = network && unavailableCapabilities[network.symbol];
-    if (network && unavailableCapability) {
-        return (
-            <Wrapper {...wrapperProps}>
-                <NetworkUnavailable capability={unavailableCapability} network={network} />
-            </Wrapper>
-        );
-    }
-
-    // Display: Network is not enabled in settings
-    if (network && !networkEnabled) {
-        const onEnableNetwork = () => {
-            props.onCancel();
+    const onEnableNetwork = () => {
+        onCancel();
+        if (network) {
             changeCoinVisibility(network.symbol, true);
-            if (app === 'wallet' && !props.noRedirect) {
+            if (app === 'wallet' && !noRedirect) {
                 // redirect to account only if added from "wallet" app
                 goto('wallet-index', {
                     symbol: network.symbol,
@@ -108,45 +106,31 @@ const AddAccount = (props: Props) => {
                     accountType: 'normal',
                 });
             }
-        };
-
-        return (
-            <Wrapper
-                {...wrapperProps}
-                actionButton={
-                    <Button variant="primary" onClick={onEnableNetwork}>
-                        <Translation
-                            id="TR_ENABLE_NETWORK_BUTTON"
-                            values={{ networkName: network.name }}
-                        />
-                    </Button>
-                }
-            />
-        );
-    }
+        }
+    };
 
     // Collect all empty accounts related to selected device and selected accountType
     const currentType = network?.accountType ?? 'normal';
-    const emptyAccounts =
-        network &&
-        accounts.filter(
-            a =>
-                a.deviceState === props.device.state &&
-                a.symbol === network.symbol &&
-                a.accountType === currentType &&
-                a.empty,
-        );
+    const emptyAccounts = network
+        ? accounts.filter(
+              a =>
+                  a.deviceState === device.state &&
+                  a.symbol === network.symbol &&
+                  a.accountType === currentType &&
+                  a.empty,
+          )
+        : [];
 
     // Collect possible accountTypes
     const accountTypes =
-        network?.networkType === 'bitcoin'
+        networkEnabled && network?.networkType === 'bitcoin'
             ? NETWORKS.filter(n => n.symbol === network.symbol)
             : undefined;
 
     const onEnableAccount = (account: Account) => {
-        props.onCancel();
+        onCancel();
         changeAccountVisibility(account);
-        if (app === 'wallet' && !props.noRedirect) {
+        if (app === 'wallet' && !noRedirect) {
             // redirect to account only if added from "wallet" app
             goto('wallet-index', {
                 symbol: account.symbol,
@@ -156,25 +140,73 @@ const AddAccount = (props: Props) => {
         }
     };
 
-    // Display: default add account window
+    const selectedNetworks = network ? [network.symbol] : [];
+
     return (
-        <Wrapper
-            {...wrapperProps}
-            accountTypes={accountTypes}
-            actionButton={
-                network &&
-                emptyAccounts && (
-                    <AddAccountButton
-                        network={network}
-                        accounts={emptyAccounts}
-                        onEnableAccount={onEnableAccount}
-                    />
-                )
-            }
+        <StyledModal
+            cancelable
+            onCancel={onCancel}
+            heading={<Translation id="MODAL_ADD_ACCOUNT_TITLE" />}
         >
-            {network && <NetworkInternal network={network} accountTypes={accountTypes} />}
-        </Wrapper>
+            <AddAccount
+                networks={network && networkEnabled ? [network] : enabledNetworks}
+                networkCanChange={!!network && networkEnabled && !networkPinned}
+                selectedNetworks={selectedNetworks}
+                unavailableCapabilities={unavailableCapabilities}
+                handleNetworkSelection={handleNetworkSelection}
+            />
+            {!networkEnabled && (
+                <EnableNetwork
+                    networks={disabledMainnetNetworks}
+                    testnetNetworks={disabledTestnetNetworks}
+                    selectedNetworks={selectedNetworks}
+                    unavailableCapabilities={unavailableCapabilities}
+                    handleNetworkSelection={handleNetworkSelection}
+                />
+            )}
+            {network && (
+                <>
+                    {accountTypes && accountTypes?.length > 1 && (
+                        <>
+                            <Title>
+                                <Translation id="TR_ACCOUNT_TYPE" />
+                            </Title>
+                            <AccountTypeSelect
+                                network={network}
+                                accountTypes={accountTypes}
+                                onSelectAccountType={setNetwork}
+                            />
+                        </>
+                    )}
+
+                    {unavailableCapability ? (
+                        <NetworkUnavailable capability={unavailableCapability} network={network} />
+                    ) : (
+                        <NetworkInternal network={network} accountTypes={accountTypes} />
+                    )}
+
+                    <Expander />
+
+                    <Actions>
+                        {networkEnabled ? (
+                            <AddAccountButton
+                                network={network}
+                                accounts={emptyAccounts}
+                                onEnableAccount={onEnableAccount}
+                            />
+                        ) : (
+                            <Button variant="primary" onClick={onEnableNetwork}>
+                                <Translation
+                                    id="TR_ENABLE_NETWORK_BUTTON"
+                                    values={{ networkName: network.name }}
+                                />
+                            </Button>
+                        )}
+                    </Actions>
+                </>
+            )}
+        </StyledModal>
     );
 };
 
-export default AddAccount;
+export default AddAccountModal;
